requirements.txt
dj-database-url
django
django-heroku
django-storages
dropbox
gunicorn
Pillow
python-dotenv
whitenoise